added test

+++ src/main/java/com/pooch/api/sms/SmsService.java
@@ -2,5 +2,5 @@ package com.pooch.api.sms;
 
 public interface SmsService {
 
-    boolean sendSMS(int countryCode, int phoneNumber, String message);
+    boolean sendSMS(int countryCode, long phoneNumber, String message);
 }
